ai_research_assistant/requirements.txt

djangorestframework
pdfminer.six
gemini
langchain
faiss-cpu
pymupdf
numpy
pandas
spacy
transformers